refactor: folder constructure
feat: 新機能
fix: バグ修正・レビュー反映
docs: ドキュメントの変更
style: コードのロジックに影響を与えない変更（ホワイトスペース、フォーマット、セミコロンの欠落など）
refactor: バグ修正でも機能追加でもないコード変更
perf: パフォーマンス向上
test: テストの追加や既存のテストの修正
chore: ビルドプロセス、またはドキュメント生成などの補助ツールやライブラリの変更
