+++ lib/pages/auth_page.dart
@@ -6,7 +6,7 @@ import '../utils/extensions/widget_ref.dart';
 import '../utils/loading.dart';
 import 'error_page.dart';
 import 'get_started/get_started_page.dart';
-import 'home_page.dart';
+import 'home/home_page.dart';
 
 /// 注意：画面遷移に [MaterialPageRoute] を採用しているので、ローディング時に重ねる
 /// [OverlayLoadingWidget] は各画面で実装が必要である。
