+++ lib/pages/result/result_page.dart
@@ -3,20 +3,20 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../features/app_user.dart';
-import '../features/member.dart';
-import '../features/result.dart';
-import '../models/member.dart';
-import '../models/result.dart';
-import '../repositories/auth/auth_repository_impl.dart';
-import '../utils/async_value_error_dialog.dart';
-import '../utils/constants/app_colors.dart';
-import '../utils/constants/measure.dart';
-import '../utils/extensions/date_time.dart';
-import '../utils/loading.dart';
-import '../utils/scaffold_messenger_service.dart';
-import '../utils/text_styles.dart';
-import '../widgets/white_app_bar.dart';
+import '../../features/app_user.dart';
+import '../../features/member.dart';
+import '../../features/result.dart';
+import '../../models/member.dart';
+import '../../models/result.dart';
+import '../../repositories/auth/auth_repository_impl.dart';
+import '../../utils/async_value_error_dialog.dart';
+import '../../utils/constants/app_colors.dart';
+import '../../utils/constants/measure.dart';
+import '../../utils/extensions/date_time.dart';
+import '../../utils/loading.dart';
+import '../../utils/scaffold_messenger_service.dart';
+import '../../utils/text_styles.dart';
+import '../../widgets/white_app_bar.dart';
 
 class ResultPage extends HookConsumerWidget {
   const ResultPage({
