+++ lib/pages/sign_up/sign_up_page.dart
@@ -11,7 +11,7 @@ import '../../utils/scaffold_messenger_service.dart';
 import '../../widgets/gradation_background.dart';
 import '../../widgets/rounded_button.dart';
 import '../../widgets/white_app_bar.dart';
-import '../home_page.dart';
+import '../home/home_page.dart';
 import 'src/sign_up.dart';
 
 class SignUpPage extends HookConsumerWidget {
