+++ lib/widgets/match_result_card.dart
@@ -7,7 +7,7 @@ import '../features/app_user.dart';
 import '../features/member.dart';
 import '../models/member.dart';
 import '../models/result.dart';
-import '../pages/result_page.dart';
+import '../pages/result/result_page.dart';
 import '../utils/constants/app_colors.dart';
 import '../utils/constants/measure.dart';
 import '../utils/extensions/date_time.dart';
